--- Frontend/todo-front-end/src/components/AuthenticationSmallPanel/AuthenticationSmallPanel.tsx
@@ -1,9 +1,0 @@
-
-
-
-export default function (){
-    return(<>
-
-
-    </>)
-}
