+++ Frontend/todo-front-end/src/components/authentication/RegisterComponent/RegisterComponent.css
